+++ features/bootstrap/CacheContext.php
@@ -1,13 +1,17 @@
 <?php
 
+namespace test\Cmp\Cache;
+
 use Behat\Behat\Context\SnippetAcceptingContext;
 use Cmp\Cache\Backend\RedisCache;
 
 /**
- * Class FeatureContext
+ * Class CacheContext
  */
-class FeatureContext implements SnippetAcceptingContext
+class CacheContext implements SnippetAcceptingContext
 {
+    use ServiceProviderTrait;
+
     /**
      * @var \Redis
      */
@@ -31,6 +35,14 @@ class FeatureContext implements SnippetAcceptingContext
         $this->redis = new \Redis();
         $this->redis->connect($_SERVER['REDIS_HOST'], 6379);
         $this->backend = new RedisCache($this->redis);
+    }
+
+    /**
+     * @return Redis
+     */
+    protected function getRedis()
+    {
+        return $this->redis;
     }
 
     /**